Extracts data from h5 files to plot in gnuplot
If given two arguments, expects the first to be the .h5 filename and the second to be a GROUP containing a Green function (or self energy, hybridisation function, etc.). Will produce ascii file with three columns: wn, re, im.

If given three arguments, expects the first to be the .h5 filename, the second to be a GROUP containing a vertex function, and the third to be a bosonic frequency index. For that bosonic frequency, will produce ascii file with four  columns: wn, wn', re, im.